fixed. LadderLine class
